chore: initial switch mock
