+++ frontend/src/components/Home/components/DefaultHomePage.js
@@ -1,0 +1,6 @@
+import React from "react";
+
+const DefaultHomePage = () => {
+  return <div>HomePage</div>;
+};
+export default DefaultHomePage;
